request validations added

+++ src/main/java/com/scopic/antiqueauction/domain/converter/AntiqueResponseConverter.java
@@ -4,11 +4,12 @@ import com.scopic.antiqueauction.domain.entity.Antique;
 import com.scopic.antiqueauction.domain.entity.PastBid;
 import com.scopic.antiqueauction.domain.response.AntiqueResponse;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 
 public class AntiqueResponseConverter {
-    public static AntiqueResponse convert(final Antique antique ,final List<BigInteger> bids,List<String> imagePaths){
+    public static AntiqueResponse convert(final Antique antique , final List<BigDecimal> bids, List<String> imagePaths){
         final AntiqueResponse antiqueResponse=new AntiqueResponse();
         antiqueResponse.setId(antique.getId());
         antiqueResponse.setName(antique.getName());
